requirements.txt
requests==2.31.0
pandas==2.3.3
finvizfinance==1.3.0
python-dotenv==1.2.2
pytz==2026.1.post1
python-telegram-bot==22.6
